WatchTower/watchtower-agent/requirements.txt
psutil==5.9.4
requests==2.28.2
PyYAML==6.0